requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data / Numerics
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0

# Logging (elige uno)
tensorboard>=2.13.0
# wandb>=0.15.0  # alternativa: descomentar si se prefiere W&B
